+++ src/Util/TerrainConverter.lua
@@ -11,26 +11,6 @@ local function FillCylinder(part, material)
   return true
 end
 
--- ! Deprecated: Use Roblox's FillWedge API.
-local function ConvertWedge(wedge, mat)
-  local x = wedge.Size.X
-  for i = 0, x do
-    local point1 = wedge.CFrame*CFrame.new(x/-2+i,wedge.Size.Y/2,wedge.Size.Z/2)
-    local point2 = wedge.CFrame*CFrame.new(x/-2+i,wedge.Size.Y/-2,wedge.Size.Z/-2)
-    local point3 = wedge.CFrame*CFrame.new(x/-2+i,wedge.Size.Y/-2,wedge.Size.Z/2)
-    local lerpGuess = (point2.p-point1.p).magnitude + (point3.p-point1.p).magnitude
-    for lerpA = 0, 1, 1/lerpGuess do
-      local p1 = point1.p:lerp(point2.p,lerpA)
-      local p2 = point3.p:lerp(point2.p,lerpA)
-      local dist = (p2-p1).magnitude
-      for lerpB = 0,1,1/dist do
-        local p = point1.p:lerp(point2.p,lerpB)
-        workspace.Terrain:FillBall(p,1,mat)
-      end
-    end
-  end
-end
-
 local function Convert(part, material)
   if part:IsA('Part') then
     if part.Shape == Enum.PartType.Block then
@@ -46,7 +26,7 @@ local function Convert(part, material)
       error('Notice.ShapeNotSupported',0)
     end
   elseif part:IsA('WedgePart') then
-    ConvertWedge(part, material)
+    workspace.Terrain:FillWedge(part.CFrame, part.Size, material)
     return true
   else
     error('Notice.ClassNotSupport',0)