+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Changed
+* Wedges now use Roblox's `FillWedge` function which allows for a full wedge shape rather than only converting the top part of the wedge.
   #### UI
   * Rewritten to use Roact.
   * Settings now has a proper toggle button reather than a checkbox.